+++ cdm_reader_mapper/__init__.py
@@ -2,16 +2,16 @@
 
 from __future__ import annotations
 
-from .cdm_mapper.map_model import map_model  # noqa
+from .cdm_mapper.mapper import map_model  # noqa
 from .cdm_mapper.properties import cdm_tables  # noqa
-from .cdm_mapper.read import read_tables  # noqa
-from .cdm_mapper.write import write_tables  # noqa
+from .cdm_mapper.reader import read_tables  # noqa
+from .cdm_mapper.writer import write_tables  # noqa
 from .core import DataBundle  # noqa
 from .data import test_data  # noqa
 from .duplicates.duplicates import DupDetect  # noqa
 from .duplicates.duplicates import duplicate_check  # noqa
-from .mdf_reader.read import read_data, read_mdf  # noqa
-from .mdf_reader.write import write_data  # noqa
+from .mdf_reader.reader import read_data, read_mdf  # noqa
+from .mdf_reader.writer import write_data  # noqa
 from .metmetpy.datetime.correct import correct as correct_datetime  # noqa
 from .metmetpy.datetime.validate import validate as validate_datetime  # noqa
 from .metmetpy.platform_type.correct import correct as correct_pt  # noqa
